@@ -1,114 +1,120 @@
 {
-  "name": "letters-social",
-  "version": "0.2.0",
-  "description": "A newsfeed application for Letters.",
-  "main": "./src/index.js",
-  "scripts": {
-    "ssr": "next",
-    "build-css": "node-sass src/ -o src/",
-    "watch-css": "npm run build-css && node-sass src/ -o src/ --watch --recursive",
-    "api:dev": "json-server db/seed/db.json -r db/routes.json -p 3500 -s db/backups",
-    "api:prod": "json-server db/seed/db.json -r db/routes.json -p ${ENDPOINT_PORT} -s db/backups",
-    "build:app": "node build/build.js",
-    "build:client": "react-scripts build",
-    "coverage": "cat ./coverage/lcov.info | coveralls",
-    "client:dev": "node run/dev.js",
-    "client:test": "jest --coverage --silent",
-    "client:test:watch": "jest --watch",
-    "build": "npm-run-all build:app build:server copy:html db:seed",
-    "build:server": "babel server/*.js --babelrc=true --ignore='node_modules, coverage, docs, dist' --compact='true' --minified --no-comments --out-dir dist",
-    "copy:html": "cp -r public/index.html dist",
-    "dev": "npm-run-all --parallel api:dev client:dev client:test:watch styles:w",
-    "db:seed": "node db/seed.js",
-    "styles:w": "node-sass src/styles/styles.scss -o public/ -w src/styles",
-    "stlyes": "node-sass src/styles/styles.scss -o public/",
-    "start:dev": "node run/dev.js",
-    "rundep": "react-scripts start",
-    "test": "npm run client:test && npm run coverage",
-    "start": "node --optimize_for_size ./server/run.js",
-    "welcome": "node run/welcome.js"
-  },
-  "jest": {
-    "testEnvironment": "jsdom"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+ssh://git@github.com/react-in-action/letters-social.git"
-  },
-  "author": "Mark Thomas <mark@ifelse.io>",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/react-in-action/letters-social/issues"
-  },
-  "engines": {
-    "node": "7.2.0"
-  },
-  "homepage": "https://github.com/react-in-action/letters-social#readme",
-  "dependencies": {
-    "autoprefixer": "6.7.4",
-    "axios": "0.15.3",
-    "bad-words": "1.5.0",
-    "bluebird": "3.4.7",
-    "body-parser": "1.16.1",
-    "bootstrap-sass": "3.3.7",
-    "chalk": "1.1.3",
-    "cheerio": "0.22.0",
-    "compression": "1.6.2",
-    "connect-history-api-fallback": "1.3.0",
-    "cookie-parser": "1.4.3",
-    "cors": "2.8.1",
-    "enroute": "1.0.1",
-    "environs": "0.4.1",
-    "enzyme": "2.7.1",
-    "eslint": "3.16.0",
-    "eslint-config-airbnb": "14.1.0",
-    "eslint-plugin-import": "2.2.0",
-    "eslint-plugin-jsx-a11y": "4.0.0",
-    "eslint-plugin-react": "6.10.0",
-    "express": "4.14.1",
-    "faker": "4.1.0",
-    "firebase": "3.6.9",
-    "flexboxgrid": "6.3.1",
-    "flow-bin": "0.39.0",
-    "helmet": "3.4.0",
-    "history": "4.5.1",
-    "hpp": "0.2.1",
-    "invariant": "2.2.2",
-    "isomorphic-fetch": "2.2.1",
-    "jest": "19.0.0",
-    "json-loader": "0.5.4",
-    "json-server": "0.9.5",
-    "loaders.css": "0.1.2",
-    "localforage": "1.5.0",
-    "lodash": "4.17.4",
-    "mkdirp": "0.5.1",
-    "moment": "2.17.1",
-    "morgan": "1.8.1",
-    "next": "2.1.1",
-    "node-sass": "4.5.0",
-    "node-uuid": "1.4.7",
-    "ora": "1.1.0",
-    "postcss-loader": "1.3.1",
-    "react": "15.4.2",
-    "react-addons-test-utils": "15.4.2",
-    "react-bootstrap": "0.30.7",
-    "react-dom": "15.5.4",
-    "react-overlays": "0.6.11",
-    "react-redux": "5.0.2",
-    "react-router-redux": "4.0.8",
-    "react-scripts": "0.9.0",
-    "react-test-renderer": "15.4.2",
-    "redux": "3.6.0",
-    "redux-logger": "2.8.1",
-    "redux-thunk": "2.2.0",
-    "response-time": "2.3.2"
-  },
-  "devDependencies": {
-    "coveralls": "2.11.16",
-    "jest-enzyme": "2.1.2",
-    "nock": "9.0.6",
-    "nyc": "10.1.2",
-    "redux-mock-store": "1.2.2",
-    "sinon": "1.17.7"
-  }
+    "name": "letters-social",
+    "version": "0.2.0",
+    "description": "A newsfeed application for Letters.",
+    "main": "./src/index.js",
+    "scripts": {
+        "build-css": "node-sass src/ -o src/",
+        "watch-css": "npm run build-css && node-sass src/ -o src/ --watch --recursive",
+        "api:dev": "json-server db/seed/db.json -r db/routes.json -p 3500 -s db/backups",
+        "api:prod": "json-server db/seed/db.json -r db/routes.json -p ${ENDPOINT_PORT} -s db/backups",
+        "build:app": "node build/build.js",
+        "build:client": "react-scripts build",
+        "coverage": "cat ./coverage/lcov.info | coveralls",
+        "client:dev": "node run/dev.js",
+        "client:test": "jest --coverage --silent",
+        "client:test:watch": "jest --watch",
+        "build": "npm-run-all build:app build:server copy:html db:seed",
+        "build:server": "babel server/*.js --babelrc=true --ignore='node_modules, coverage, docs, dist' --compact='true' --minified --no-comments --out-dir dist",
+        "copy:html": "cp -r public/index.html dist",
+        "dev": "npm-run-all --parallel api:dev client:dev client:test:watch styles:w",
+        "db:seed": "node db/seed.js",
+        "styles:w": "node-sass src/styles/styles.scss -o public/ -w src/styles",
+        "stlyes": "node-sass src/styles/styles.scss -o public/",
+        "start:dev": "node run/dev.js",
+        "rundep": "react-scripts start",
+        "test": "npm run client:test && npm run coverage",
+        "start": "node --optimize_for_size ./server/run.js",
+        "welcome": "node run/welcome.js"
+    },
+    "repository": {
+        "type": "git",
+        "url": "git+ssh://git@github.com/react-in-action/letters-social.git"
+    },
+    "author": "Mark Thomas <mark@ifelse.io>",
+    "license": "MIT",
+    "bugs": {
+        "url": "https://github.com/react-in-action/letters-social/issues"
+    },
+    "engines": {
+        "node": "7.2.0"
+    },
+    "jest": {
+        "testEnvironment": "jsdom"
+    },
+    "homepage": "https://github.com/react-in-action/letters-social#readme",
+    "dependencies": {
+        "autoprefixer": "6.7.7",
+        "axios": "0.16.1",
+        "bad-words": "1.5.1",
+        "bluebird": "3.5.0",
+        "body-parser": "1.17.1",
+        "bootstrap-sass": "3.3.7",
+        "browser-sync": "2.18.8",
+        "chalk": "1.1.3",
+        "cheerio": "0.22.0",
+        "compression": "1.6.2",
+        "connect-history-api-fallback": "1.3.0",
+        "cookie-parser": "1.4.3",
+        "cors": "2.8.3",
+        "enroute": "1.0.1",
+        "environs": "0.4.1",
+        "enzyme": "2.8.2",
+        "eslint": "3.19.0",
+        "eslint-plugin-jsx-a11y": "4.0.0",
+        "express": "4.15.2",
+        "faker": "4.1.0",
+        "firebase": "3.8.0",
+        "flexboxgrid": "6.3.1",
+        "helmet": "3.5.0",
+        "history": "4.6.1",
+        "hpp": "0.2.2",
+        "invariant": "2.2.2",
+        "isomorphic-fetch": "2.2.1",
+        "json-server": "0.9.6",
+        "loaders.css": "0.1.2",
+        "localforage": "1.5.0",
+        "lodash": "4.17.4",
+        "moment": "2.18.1",
+        "morgan": "1.8.1",
+        "node-uuid": "1.4.8",
+        "ora": "1.2.0",
+        "react": "15.5.4",
+        "react-bootstrap": "0.30.9",
+        "react-dom": "15.5.4",
+        "react-overlays": "0.6.12",
+        "react-redux": "5.0.4",
+        "react-router-redux": "4.0.8",
+        "redux": "3.6.0",
+        "redux-logger": "2.8.1",
+        "redux-thunk": "2.2.0",
+        "response-time": "2.3.2"
+    },
+    "devDependencies": {
+        "babel-jest": "19.0.0",
+        "babel-plugin-transform-class-properties": "^6.24.1",
+        "babel-preset-es2015": "6.24.1",
+        "babel-preset-react": "6.24.1",
+        "coveralls": "2.13.0",
+        "eslint-config-airbnb": "14.1.0",
+        "eslint-plugin-import": "2.2.0",
+        "eslint-plugin-react": "6.10.3",
+        "html-webpack-plugin": "^2.28.0",
+        "jest": "19.0.2",
+        "jest-enzyme": "3.0.1",
+        "mkdirp": "0.5.1",
+        "nock": "9.0.13",
+        "node-sass": "4.5.2",
+        "nyc": "10.2.0",
+        "postcss-loader": "1.3.3",
+        "react-addons-test-utils": "15.5.1",
+        "react-test-renderer": "15.5.4",
+        "redux-mock-store": "1.2.3",
+        "regenerator-runtime": "0.10.3",
+        "sass-loader": "6.0.3",
+        "sinon": "2.1.0",
+        "webpack": "2.4.1",
+        "webpack-dev-middleware": "1.10.1",
+        "webpack-hot-middleware": "2.18.0"
+    }
 }
