@@ -1,4 +1,5 @@
-import React, { PropTypes } from 'react';
+import PropTypes from 'prop-types';
+import React from 'react';
 
 class CreateComment extends React.Component {
     constructor(props) {
@@ -8,17 +9,17 @@ class CreateComment extends React.Component {
         };
         this.handleCommentUpdate = this.handleCommentUpdate.bind(this);
     }
+    handleCommentUpdate(event) {
+        this.setState({
+            comment: event.target.value
+        });
+    }
     handleSubmit(event) {
         event.preventDefault();
         const { threadID, handleSubmit } = this.props;
         console.log('Submitting');
         console.log(event.target);
         handleSubmit(threadID);
-    }
-    handleCommentUpdate(event) {
-        this.setState({
-            comment: event.target.value
-        });
     }
     render() {
         return (
