@@ -1,18 +1,17 @@
-import { PropTypes, Children, Component, cloneElement } from 'react';
+import PropTypes from 'prop-types';
+import { Children, cloneElement } from 'react';
 import { navigate } from '../../history/';
 
-class Link extends Component {
-    static propTypes = {
-        to: PropTypes.string,
-        children: PropTypes.node
-    };
-
-    render() {
-        const { to, children } = this.props;
-        return cloneElement(Children.only(children), {
-            onClick: () => navigate(to)
-        });
-    }
+function Link(props) {
+    const { to, children } = props;
+    return cloneElement(Children.only(children), {
+        onClick: () => navigate(to)
+    });
 }
 
-export { Link };
+Link.propTypes = {
+    to: PropTypes.string,
+    children: PropTypes.node
+};
+
+export default Link;
