@@ -1,6 +1,8 @@
 import createHistory from 'history/createBrowserHistory';
 
-const history = createHistory();
+const history = typeof window !== 'undefined'
+    ? createHistory()
+    : { push: () => {} };
 const navigate = to => history.push(to);
 
 export { history, navigate };
