@@ -1,7 +1,7 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 
-import Navigation from './navbar';
+import Navigation from '../../../src/components/nav/Navbar';
 
 describe('<Navigation/>', () => {
     it('should render correctly', () => {
