@@ -1,12 +1,12 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 
-import { Login } from './Login';
+import Login from '../../src/pages/login';
 
 describe('<Login/>', () => {
-  it('Login should render correctly', () => {
-    const component = renderer.create(<Login />);
-    const tree = component.toJSON();
-    expect(tree).toMatchSnapshot();
-  });
+    it('Login should render correctly', () => {
+        const component = renderer.create(<Login />);
+        const tree = component.toJSON();
+        expect(tree).toMatchSnapshot();
+    });
 });
